Hide glm::outerProduct_trait from SWIG
SWIG runs into a syntax error with outerProduct_trait, because it has
a template within a template.

This only seems to be used by glm::outerProduct(). If we ever need that
function from Lua, we'll have to manually (or using a macro) declare
all of the possible functions, of which there seem to be 9 versions.